Time: 140 ms (14.08%), Space: 32.2 MB (76.96%) - LeetHub

+++ 1465-maximum-area-of-a-piece-of-cake-after-horizontal-and-vertical-cuts/1465-maximum-area-of-a-piece-of-cake-after-horizontal-and-vertical-cuts.cpp
@@ -34,8 +34,8 @@ public:
         
 //         return (height*length) % 1000000007;
         
-        long length=solve(horizontalCuts,h);
-        long height=solve(verticalCuts,w);
+        long long length=solve(horizontalCuts,h);
+        long long height=solve(verticalCuts,w);
         
         return (length*height)%1000000007;
     }
